Parent groups added
Now when registering groups, you can specify the parent group.
The parent group is used when checking the access of one label to another: if, when comparing, the label groups are
not equal, then the parent label group to which access is requested (object label) is compared with the subject label
group, and so on until the groups are equal or the parent group of the label of the object will not be -1 (absence of
the parent group).