Added automatic label ID generation
Now possible not explicitly specify the ID of the label of object, it will be created  based on the recorded
components of the label.
When recording identical labels of object without specifying an identifier, only one mark will be recorded,
because with the same label components, the same ID is generated.